insurance/requirement.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
buildurl==4.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
contextvars==2.4
dataclasses==0.8
Django==3.2.11
django-ajax-datatable==4.4.0
djangorestframework==3.13.1
dnspython==2.2.0
eventlet==0.33.0
frozenlist==1.2.0
gevent==21.12.0
greenlet==1.1.2
grequests==0.6.0
h11==0.12.0
httpcore==0.14.5
httpx==0.22.0
idna==3.3
idna-ssl==1.1.0
immutables==0.16
importlib-metadata==4.8.3
Markdown==3.3.6
multidict==5.2.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
